Количества в BigDecimal вместо Double
На Double дробные дозы не сходились: после трёх списаний по 1/3 остаток равнялся не нулю, а
величине порядка 1e-16. Отсюда два наблюдаемых дефекта — кончившийся препарат не удалялся, и
план оставался висеть с ничтожным остатком, потому что обе проверки были точными сравнениями
с 0.0. Инвариант «сумма планов не больше остатка» при этом тихо нарушался.

Drug.quantity, Drug.totalPlannedAmount и Using.plannedAmount переведены в BigDecimal,
колонки — в numeric(19,6). Скрипт для существующей БД: db/manual/001_decimal_quantities.sql,
применяется с остановленным сервером, потому что в проде ddl-auto=validate.

Ключевое про сравнения. У BigDecimal equals учитывает scale, поэтому 0.000000 не равен ZERO,
и наивный перенос увёз бы тот же баг из Double в BigDecimal, только менее заметно. Введены
явные isZero/isPositive и toQuantityScale, сравнения на порядок идут через compareTo (в
Kotlin это обычные > и <=, они корректны). Деление коэффициента сжатия — с явными scale и
режимом округления, иначе BigDecimal бросает исключение на бесконечной дроби.

Остаток округления при пропорциональном уменьшении планов отдаётся одному плану — самому
большому, чтобы поправка не загнала его в минус, при равенстве выбор детерминирован по
userId. Без этого сумма округлённых планов не сходилась с остатком, а именно из таких копеек
и вырастало расхождение.

Попутно найден и исправлен латентный дефект, который вскрыл новый тест: удаление
кончившегося препарата падало с нарушением внешнего ключа. Каскад CascadeType.ALL объявлен,
но идёт по ленивой коллекции drug.usings, которая в этом пути не синхронизирована —
createTreatmentPlan сохраняет план через usingRepository.save и в коллекцию его не добавляет.
На Double эта ветка практически не исполнялась, поэтому дефект не проявлялся. Планы теперь
удаляются запросом с явным flush перед удалением препарата.

В тестах: qty() для литералов через BigDecimal.valueOf (конструктор от Double дал бы точное
двоичное представление и не сошёлся бы ни с базой, ни с ожиданием) и assertQty для сравнения
по значению без учёта scale. Заодно исчезли 32 сравнения с допуском 0.001 — с точными
числами допуск не нужен.

open-api.yaml перегенерирован: количества стали "type: number" без "format: double".
Снапшот-тест сам это и поймал, для чего и вводился.

Прогон: 202 теста, 0 падений, интеграционные — на настоящем Postgres.

+++ src/MedAppServer/db/schema.sql
@@ -49,7 +49,7 @@ CREATE TABLE user_drugs
 (
     id            uuid             NOT NULL,
     name          varchar(300)     NOT NULL,
-    quantity      double precision NOT NULL,
+    quantity      numeric(19, 6)   NOT NULL,
     quantity_unit varchar(50)      NOT NULL,
     form_type     varchar(100),
     category      varchar(200),
@@ -71,7 +71,7 @@ CREATE TABLE usings
 (
     user_id        uuid                        NOT NULL,
     drug_id        uuid                        NOT NULL,
-    planned_amount double precision            NOT NULL,
+    planned_amount numeric(19, 6)              NOT NULL,
     created_at     timestamp(6) with time zone NOT NULL,
     last_modified  timestamp(6) with time zone NOT NULL,
